+++ lib/service/test/definition.go
@@ -3,6 +3,7 @@ package test
 import (
 	"fmt"
 
+	"github.com/Jeffail/benthos/v3/lib/log"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -24,10 +25,20 @@ func ExampleDefinition() Definition {
 
 //------------------------------------------------------------------------------
 
+// ExecuteWithLogger attempts to run a test definition on a target config file,
+// with a logger. Returns an array of test failures or an error.
+func (d Definition) ExecuteWithLogger(filepath string, logger log.Modular) ([]CaseFailure, error) {
+	return d.execute(filepath, logger)
+}
+
 // Execute attempts to run a test definition on a target config file. Returns
 // an array of test failures or an error.
 func (d Definition) Execute(filepath string) ([]CaseFailure, error) {
-	procsProvider := NewProcessorsProvider(filepath)
+	return d.execute(filepath, log.Noop())
+}
+
+func (d Definition) execute(filepath string, logger log.Modular) ([]CaseFailure, error) {
+	procsProvider := NewProcessorsProvider(filepath, OptProcessorsProviderSetLogger(logger))
 	if d.Parallel {
 		// Warm the cache of processor configs.
 		for _, c := range d.Cases {