+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Flag `log` added to `test` sub command to allow logging during tests.
+
 ## 3.14.0 - 2020-05-17
 
 ### Added